requirements.txt
setuptools
wheel
python-dotenv
pre-commit
custom-python-logger
retrying
elasticsearch<9
pandas
psycopg2-binary
pytest
